--- src/main/java/core/mvc/JspView.java
@@ -1,21 +1,0 @@
-package core.mvc;
-
-import javax.servlet.RequestDispatcher;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.util.Map;
-
-public class JspView implements View {
-    private final String viewName;
-
-    public JspView(String viewName) {
-        this.viewName = viewName;
-    }
-
-    @Override
-    public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        RequestDispatcher rd = request.getRequestDispatcher(viewName);
-        model.forEach(request::setAttribute);
-        rd.forward(request, response);
-    }
-}
