+++ src/main/java/core/mvc/tobe/Handler.java
@@ -1,0 +1,10 @@
+package core.mvc;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public interface Handler {
+    ModelAndView handle(HttpServletRequest request, HttpServletResponse response);
+
+    boolean isSupport(HttpServletRequest req);
+}
