+++ src/main/java/core/mvc/view/RedirectView.java
@@ -1,0 +1,22 @@
+package core.mvc;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.Map;
+
+public class RedirectView implements View {
+    private String viewName;
+
+    public RedirectView(String viewName) {
+        this.viewName = viewName;
+    }
+
+    public String getViewName() {
+        return viewName;
+    }
+
+    @Override
+    public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception {
+        response.sendRedirect(viewName);
+    }
+}
